pyproject.toml
[tool.poetry]
name = "TibiaPilotNG"
version = "0.1.0"
description = ""
authors = ["Paulo Lacerda <paulorl2810@gmail.com>", "lucasmonstro <joao.galiano.silva@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
dxcam = {git = "https://github.com/paulordyl/DXcam"}
cityhash = "^0.4.7"
mypy = "^1.1.1"
nptyping = "^2.5.0"
numba = "^0.58.1"
numpy = "1.26.2"
opencv-python = "^4.6.0.66"
pillow = "^9.1.1"
pyautogui = "^0.9.53"
pygetwindow = "^0.0.9"
python = "3.11.7"
pywin32 = "^306"
scipy = "^1.9.1"
tcod = "^13.8.1"
tinydb = "^4.8.0"
customtkinter = "^5.2.2"
requests = "^2.31.0"
pyserial = "^3.5"
scikit-image = "^0.22.0"
pytesseract = "^0.3.10"

[tool.poetry.group.test.dependencies]
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
pytest-mock = "^3.8.2"

[tool.poetry.group.dev.dependencies]
pyinstaller = "^5.12.0"
taskipy = "^1.10.3"

[tool.taskipy.tasks]
test = "python -m pytest ."

[tool.pytest.ini_options]
filterwarnings = [
	"ignore:.*:DeprecationWarning:nptyping\\.typing_",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"